+++ src/lib/services/contentful/mappers/ArticleMapper.ts
@@ -1,8 +1,7 @@
-import type { ContentfulArticle } from '$lib/services/contentful/types/b3.contentful';
-import type { ContentfulIFrameEntry } from '$lib/services/contentful/types/b3.contentful';
 import type { Options } from '@contentful/rich-text-html-renderer';
 import { documentToHtmlString } from '@contentful/rich-text-html-renderer';
 import { BLOCKS } from '@contentful/rich-text-types';
+import type { ContentfulArticle, ContentfulIFrameEntry } from '$lib/services/contentful/types';
 
 export class ArticleMapper {
 	mapContenfulArticleToInternal(src: ContentfulArticle): Article {
