+++ src/lib/services/contentful/types/ContentfulMenuItem.ts
@@ -1,0 +1,4 @@
+export type ContentfulMenuItem = {
+	url: string;
+	text: string;
+};
