+++ src/lib/services/contentful/types/ContentfulTopBanner.ts
@@ -1,0 +1,13 @@
+export type ContentfulTopBanner = {
+	topBannerTitle: string;
+	topBannerSubtitle: string;
+	topBannerImage: {
+		url: string;
+		title: string;
+		description: string;
+	};
+	primaryLink: string | null;
+	primaryLinkText: string | null;
+	secondaryLink: string | null;
+	secondaryLinkText: string | null;
+};
