+++ src/lib/services/contentful/types/ContentfulVideo.ts
@@ -1,0 +1,5 @@
+export type ContentfulVideo = {
+	id: number;
+	title?: string;
+	description?: string;
+};
