+++ src/lib/services/contentful/types/index.ts
@@ -1,0 +1,4 @@
+export type { ContentfulArticle } from './ContentfulArticle';
+export type { ContentfulMenuItem } from './ContentfulMenuItem';
+export type { ContentfulTopBanner } from './ContentfulTopBanner';
+export type { ContentfulVideo } from './ContentfulVideo';
