+++ src/lib/services/index.ts
@@ -1,18 +1,16 @@
-import type { Article, TopBanner, Video } from '$lib/types/b3.pagedata';
-
-interface IMenuService {
+export interface IMenuService {
 	getMenuItems(): Promise<MenuItem[]>;
 }
 
-interface IArticleService {
+export interface IArticleService {
 	getArticles(): Promise<Article[]>;
 	getArticleById(id: string): Promise<Article>;
 }
 
-interface IHomePageService {
+export interface IHomePageService {
 	getTopBanner(): Promise<TopBanner>;
 }
 
-interface IVideoService {
+export interface IVideoService {
 	getFrontPageVideos(): Promise<Video[]>;
 }
