+++ src/lib/services/vimeo/types/VimeoVideoMetadata.ts
@@ -1,0 +1,9 @@
+export type VimeoVideoMetadata = {
+	files: VimeoVideoFile[];
+};
+
+type VimeoVideoFile = {
+	type: string;
+	quality: string;
+	link: string;
+};
