--- src/lib/services/vimeo/types/b3.vimeo.d.ts
@@ -1,17 +1,0 @@
-interface IVimeoClient {
-	get<T>(path: string): Promise<T>;
-}
-
-interface IVimeoVideoMetadataService {
-	getVideoMetadata(id: number): Promise<VimeoVideoMetadata>;
-}
-
-type VimeoVideoMetadata = {
-	files: VimeoVideoFile[];
-};
-
-type VimeoVideoFile = {
-	type: string;
-	quality: string;
-	link: string;
-};
