+++ src/lib/services/vimeo/types/index.ts
@@ -1,0 +1,1 @@
+export type { VimeoVideoMetadata } from './VimeoVideoMetadata';
