+++ src/lib/types/ArticlePageData.ts
@@ -1,0 +1,7 @@
+import type { Article, MenuItem } from '$lib/types';
+
+export type ArticlePageData = {
+	article: Article;
+	articles: Article[];
+	menuItems: MenuItem[];
+};
