+++ src/lib/types/HomePageData.ts
@@ -1,0 +1,8 @@
+import type { Video, MenuItem, TopBanner, Article } from '$lib/types';
+
+export type HomePageData = {
+	topBanner: TopBanner;
+	articles: Article[];
+	menuItems: MenuItem[];
+	videos: Video[];
+};
