+++ src/lib/types/MenuItem.ts
@@ -1,0 +1,4 @@
+export type MenuItem = {
+	url: string;
+	text: string;
+};
