+++ src/lib/types/TopBanner.ts
@@ -1,0 +1,13 @@
+export type TopBanner = {
+	title: string;
+	subtitle: string;
+	image: {
+		url: string;
+		title: string;
+		description: string;
+	};
+	primaryLinkUrl: string;
+	primaryLinkText: string;
+	secondaryLinkUrl: string;
+	secondaryLinkText: string;
+};
