+++ src/lib/types/Video.ts
@@ -1,0 +1,5 @@
+export type Video = {
+	id: number;
+	title?: string;
+	description?: string;
+};
