+++ src/lib/types/index.ts
@@ -1,0 +1,6 @@
+export type { HomePageData } from './HomePageData';
+export type { ArticlePageData } from './ArticlePageData';
+export type { Article } from './Article';
+export type { MenuItem } from './MenuItem';
+export type { TopBanner } from './TopBanner';
+export type { Video } from './Video';
